fix lookup for indexes of a pointer or interface
Lookup did not handle correctly the indexes when the parent was an
interface or a pointer, as getValueByName was called recursively
but the key had been overwritten without the index. Because of this,
it was treated as if there was no index.

`TestAggregableLookupString_Complex` was incorrect, it is now fixed.

A new `ErrorIndexOutOfBounds` is introduced to avoid a panic when
an index is requested but the slice is smaller than that.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>