fix panic in lookup for multiple maps
When the lookup results in multiple maps there was an issue in the
`mergeValue` function and maps were marked as mergeable, but then
handled as if they were slices. Because of that, a panic was raised.

This PR fixes that behavior by making them not mergeable instead
of throwing a panic.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>